authz: wire up permission-backend in create-app
This is not something we'd look to merge into create-app - eventually
the authorization-framework should be smart enough to short-circuit to
ALLOW responses when no permission-backend is set up. However not having
it set up in create-app is making the E2E tests fail in CI, which is a
bit distracting for the RFC, and could also hide legitimate problems
with the code.

Signed-off-by: Mike Lewis <mtlewis@users.noreply.github.com>